Move branch_operator.py to branch.py
Part of #11178 for the branch_operator
