+++ airflow/operators/branch_operator.py
@@ -15,38 +15,13 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""Branching operators"""
+"""This module is deprecated. Please use `airflow.operators.branch`."""
 
-from typing import Dict, Iterable, Union
+import warnings
 
-from airflow.models import BaseOperator
-from airflow.models.skipmixin import SkipMixin
+# pylint: disable=unused-import
+from airflow.operators.branch import BaseBranchOperator  # noqa
 
-
-class BaseBranchOperator(BaseOperator, SkipMixin):
-    """
-    This is a base class for creating operators with branching functionality,
-    similarly to BranchPythonOperator.
-
-    Users should subclass this operator and implement the function
-    `choose_branch(self, context)`. This should run whatever business logic
-    is needed to determine the branch, and return either the task_id for
-    a single task (as a str) or a list of task_ids.
-
-    The operator will continue with the returned task_id(s), and all other
-    tasks directly downstream of this operator will be skipped.
-    """
-
-    def choose_branch(self, context: Dict) -> Union[str, Iterable[str]]:
-        """
-        Subclasses should implement this, running whatever logic is
-        necessary to choose a branch and returning a task_id or list of
-        task_ids.
-
-        :param context: Context dictionary as passed to execute()
-        :type context: dict
-        """
-        raise NotImplementedError
-
-    def execute(self, context: Dict):
-        self.skip_all_except(context['ti'], self.choose_branch(context))
+warnings.warn(
+    "This module is deprecated. Please use `airflow.operators.branch`.", DeprecationWarning, stacklevel=2
+)
